Capture a plethora of information about people, vs just email
We'll be capturing most of this data in the WSU People directory
and will one day need a way to pull it in automatically. Until then
we should also allow for it to be entered manually. The only
unfortunate part of this will be displaying it manually on the front
end as well, but we'll get there.